remove access to the storage from the endpoint and simplify the endpoint sending emails
